+++ pgjdbc/src/test/java/org/postgresql/jdbcurlresolver/PgServiceConfParserTest.java
@@ -7,7 +7,7 @@ package org.postgresql.jdbcurlresolver;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
-import static org.junit.jupiter.api.Assertions.assertNull;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.postgresql.PGEnvironment;
@@ -32,20 +32,20 @@ import java.util.Properties;
  * @author Marek Läll
  */
 @ExtendWith(SystemStubsExtension.class)
-class PgServiceConfParserTest {
+public class PgServiceConfParserTest {
 
   // "org.postgresql.pgservicefile" : missing
   // "PGSERVICEFILE"                : missing
   // ".pg_service.conf"             : missing
   // "PGSYSCONFDIR"                 : missing
   @Test
-  void pgService11() throws Exception {
+  public void pgService11() throws Exception {
     Resources.with(
         new EnvironmentVariables(PGEnvironment.PGSERVICEFILE.getName(), "", PGEnvironment.PGSYSCONFDIR.getName(), ""),
         new SystemProperties(PGEnvironment.ORG_POSTGRESQL_PGSERVICEFILE.getName(), "", "user.home", "/tmp/dir-non-existent")
     ).execute(() -> {
-      Properties result = PgServiceConfParser.getServiceProperties("service-non-existent");
-      assertNull(result);
+      JdbcUrlResolverFatalException exception = assertThrows(JdbcUrlResolverFatalException.class, () -> PgServiceConfParser.getServiceProperties("service-non-existent"));
+      assertEquals("Resource file [.pg_service.conf] not found", exception.getMessage());
     });
   }
 
@@ -55,16 +55,16 @@ class PgServiceConfParserTest {
   // "PGSYSCONFDIR"                 : exist
   // <service>                      : missing
   @Test
-  void pgService21() throws Exception {
+  public void pgService21() throws Exception {
     URL urlPath = getClass().getResource("/pg_service");
     assertNotNull(urlPath);
     Resources.with(
         new EnvironmentVariables(PGEnvironment.PGSERVICEFILE.getName(), "", PGEnvironment.PGSYSCONFDIR.getName(), urlPath.getPath()),
         new SystemProperties(PGEnvironment.ORG_POSTGRESQL_PGSERVICEFILE.getName(), "", "user.home", "/tmp/dir-non-existent")
     ).execute(() -> {
-      Properties result = PgServiceConfParser.getServiceProperties("service-non-existent");
-      assertNull(result);
-      result = PgServiceConfParser.getServiceProperties("empty-service1");
+      JdbcUrlResolverFatalException exception = assertThrows(JdbcUrlResolverFatalException.class, () -> PgServiceConfParser.getServiceProperties("service-non-existent"));
+      assertEquals("Definition of service [service-non-existent] not found", exception.getMessage());
+      Properties result = PgServiceConfParser.getServiceProperties("empty-service1");
       assertNotNull(result);
       assertTrue(result.isEmpty());
     });
@@ -76,7 +76,7 @@ class PgServiceConfParserTest {
   // "PGSYSCONFDIR"                 : exist
   // <service>                      : exist
   @Test
-  void pgService22() throws Exception {
+  public void pgService22() throws Exception {
     URL urlPath = getClass().getResource("/pg_service");
     assertNotNull(urlPath);
     Resources.with(
@@ -85,10 +85,10 @@ class PgServiceConfParserTest {
     ).execute(() -> {
       Properties result = PgServiceConfParser.getServiceProperties("test-service1");
       assertNotNull(result);
-      assertEquals("test_dbname", PGProperty.DBNAME.get(result));
-      assertEquals("global-test-host.test.net", PGProperty.HOST.get(result));
-      assertEquals("5433", PGProperty.PORT.get(result));
-      assertEquals("admin", result.get("user"));
+      assertEquals("test_dbname", PGProperty.DBNAME.get2(result));
+      assertEquals("global-test-host.test.net", PGProperty.HOST.get2(result));
+      assertEquals("5433", PGProperty.PORT.get2(result));
+      assertEquals("admin", PGProperty.USER.get2(result));
       assertEquals(4, result.size());
     });
   }
@@ -99,14 +99,14 @@ class PgServiceConfParserTest {
   // "PGSYSCONFDIR"                 : exist - but file itself is missing
   // <service>                      : exist
   @Test
-  void pgService23() throws Exception {
+  public void pgService23() throws Exception {
     String nonExistingDir = "non-existing-dir";
     Resources.with(
         new EnvironmentVariables(PGEnvironment.PGSERVICEFILE.getName(), "", PGEnvironment.PGSYSCONFDIR.getName(), nonExistingDir),
         new SystemProperties(PGEnvironment.ORG_POSTGRESQL_PGSERVICEFILE.getName(), "", "user.home", "/tmp/dir-non-existent")
     ).execute(() -> {
-      Properties result = PgServiceConfParser.getServiceProperties("test-service1");
-      assertNull(result);
+      JdbcUrlResolverFatalException exception = assertThrows(JdbcUrlResolverFatalException.class, () -> PgServiceConfParser.getServiceProperties("test-service1"));
+      assertEquals("Failed to handle resource [non-existing-dir/pg_service.conf] with error [non-existing-dir/pg_service.conf (No such file or directory)]", exception.getMessage());
     });
   }
 
@@ -117,16 +117,16 @@ class PgServiceConfParserTest {
   // "PGSYSCONFDIR"                 : exist
   // <service>                      : missing
   @Test
-  void pgService31() throws Exception {
+  public void pgService31() throws Exception {
     URL urlPath = getClass().getResource("/pg_service");
     assertNotNull(urlPath);
     Resources.with(
         new EnvironmentVariables(PGEnvironment.PGSERVICEFILE.getName(), "", PGEnvironment.PGSYSCONFDIR.getName(), urlPath.getPath()),
         new SystemProperties(PGEnvironment.ORG_POSTGRESQL_PGSERVICEFILE.getName(), "", "user.home", urlPath.getPath())
     ).execute(() -> {
-      Properties result = PgServiceConfParser.getServiceProperties("service-non-existent");
-      assertNull(result);
-      result = PgServiceConfParser.getServiceProperties("empty-service1");
+      JdbcUrlResolverFatalException exception = assertThrows(JdbcUrlResolverFatalException.class, () -> PgServiceConfParser.getServiceProperties("service-non-existent"));
+      assertEquals("Definition of service [service-non-existent] not found", exception.getMessage());
+      Properties result = PgServiceConfParser.getServiceProperties("empty-service1");
       assertNotNull(result);
       assertTrue(result.isEmpty());
     });
@@ -138,7 +138,7 @@ class PgServiceConfParserTest {
   // "PGSYSCONFDIR"                 : exist
   // <service>                      : exist
   @Test
-  void pgService32() throws Exception {
+  public void pgService32() throws Exception {
     URL urlPath = getClass().getResource("/pg_service");
     assertNotNull(urlPath);
     Resources.with(
@@ -147,10 +147,10 @@ class PgServiceConfParserTest {
     ).execute(() -> {
       Properties result = PgServiceConfParser.getServiceProperties("test-service1");
       assertNotNull(result);
-      assertEquals(" test_dbname", PGProperty.DBNAME.get(result));
-      assertEquals("local-test-host.test.net", PGProperty.HOST.get(result));
-      assertEquals("5433", PGProperty.PORT.get(result));
-      assertEquals("admin", result.get("user"));
+      assertEquals(" test_dbname", PGProperty.DBNAME.get2(result));
+      assertEquals("local-test-host.test.net", PGProperty.HOST.get2(result));
+      assertEquals("5433", PGProperty.PORT.get2(result));
+      assertEquals("admin", PGProperty.USER.get2(result));
       assertEquals(4, result.size());
     });
   }
@@ -162,7 +162,7 @@ class PgServiceConfParserTest {
   // "PGSYSCONFDIR"                 : exist
   // <service>                      : missing
   @Test
-  void pgService41() throws Exception {
+  public void pgService41() throws Exception {
     URL urlPath = getClass().getResource("/pg_service");
     assertNotNull(urlPath);
     URL urlFileEnv = getClass().getResource("/pg_service/pgservicefileEnv.conf");
@@ -171,9 +171,9 @@ class PgServiceConfParserTest {
         new EnvironmentVariables(PGEnvironment.PGSERVICEFILE.getName(), urlFileEnv.getFile(), PGEnvironment.PGSYSCONFDIR.getName(), urlPath.getPath()),
         new SystemProperties(PGEnvironment.ORG_POSTGRESQL_PGSERVICEFILE.getName(), "", "user.home", urlPath.getPath())
     ).execute(() -> {
-      Properties result = PgServiceConfParser.getServiceProperties("service-non-existent");
-      assertNull(result);
-      result = PgServiceConfParser.getServiceProperties("empty-service1");
+      JdbcUrlResolverFatalException exception = assertThrows(JdbcUrlResolverFatalException.class, () -> PgServiceConfParser.getServiceProperties("service-non-existent"));
+      assertEquals("Definition of service [service-non-existent] not found", exception.getMessage());
+      Properties result = PgServiceConfParser.getServiceProperties("empty-service1");
       assertNotNull(result);
       assertTrue(result.isEmpty());
     });
@@ -185,7 +185,7 @@ class PgServiceConfParserTest {
   // "PGSYSCONFDIR"                 : exist
   // <service>                      : exist
   @Test
-  void pgService42() throws Exception {
+  public void pgService42() throws Exception {
     URL urlPath = getClass().getResource("/pg_service");
     assertNotNull(urlPath);
     URL urlFileEnv = getClass().getResource("/pg_service/pgservicefileEnv.conf");
@@ -196,11 +196,11 @@ class PgServiceConfParserTest {
     ).execute(() -> {
       Properties result = PgServiceConfParser.getServiceProperties("test-service1");
       assertNotNull(result);
-      assertEquals("test_dbname", PGProperty.DBNAME.get(result));
-      assertEquals("pgservicefileEnv-test-host.test.net", PGProperty.HOST.get(result));
-      assertEquals("5433", PGProperty.PORT.get(result));
-      assertEquals("admin", result.get("user"));
-      assertEquals("disable", result.get("sslmode"));
+      assertEquals("test_dbname", PGProperty.DBNAME.get2(result));
+      assertEquals("pgservicefileEnv-test-host.test.net", PGProperty.HOST.get2(result));
+      assertEquals("5433", PGProperty.PORT.get2(result));
+      assertEquals("admin", PGProperty.USER.get2(result));
+      assertEquals("disable", PGProperty.SSL_MODE.get2(result));
       assertEquals(5, result.size());
     });
   }
@@ -211,7 +211,7 @@ class PgServiceConfParserTest {
   // "PGSYSCONFDIR"                 : exist
   // <service>                      : exist
   @Test
-  void pgService43() throws Exception {
+  public void pgService43() throws Exception {
     URL urlPath = getClass().getResource("/pg_service");
     assertNotNull(urlPath);
     String nonExistingFile = "non-existing-file.conf";
@@ -219,8 +219,8 @@ class PgServiceConfParserTest {
         new EnvironmentVariables(PGEnvironment.PGSERVICEFILE.getName(), nonExistingFile, PGEnvironment.PGSYSCONFDIR.getName(), urlPath.getPath()),
         new SystemProperties(PGEnvironment.ORG_POSTGRESQL_PGSERVICEFILE.getName(), "", "user.home", urlPath.getPath())
     ).execute(() -> {
-      Properties result = PgServiceConfParser.getServiceProperties("test-service1");
-      assertNull(result);
+      JdbcUrlResolverFatalException exception = assertThrows(JdbcUrlResolverFatalException.class, () -> PgServiceConfParser.getServiceProperties("test-service1"));
+      assertEquals("Failed to handle resource [non-existing-file.conf] with error [non-existing-file.conf (No such file or directory)]", exception.getMessage());
     });
   }
 
@@ -231,7 +231,7 @@ class PgServiceConfParserTest {
   // "PGSYSCONFDIR"                 : exist
   // <service>                      : missing
   @Test
-  void pgService51() throws Exception {
+  public void pgService51() throws Exception {
     URL urlPath = getClass().getResource("/pg_service");
     assertNotNull(urlPath);
     URL urlFileEnv = getClass().getResource("/pg_service/pgservicefileEnv.conf");
@@ -242,9 +242,9 @@ class PgServiceConfParserTest {
         new EnvironmentVariables(PGEnvironment.PGSERVICEFILE.getName(), urlFileEnv.getFile(), PGEnvironment.PGSYSCONFDIR.getName(), urlPath.getPath()),
         new SystemProperties(PGEnvironment.ORG_POSTGRESQL_PGSERVICEFILE.getName(), urlFileProps.getFile(), "user.home", urlPath.getPath())
     ).execute(() -> {
-      Properties result = PgServiceConfParser.getServiceProperties("service-non-existent");
-      assertNull(result);
-      result = PgServiceConfParser.getServiceProperties("empty-service1");
+      JdbcUrlResolverFatalException exception = assertThrows(JdbcUrlResolverFatalException.class, () -> PgServiceConfParser.getServiceProperties("service-non-existent"));
+      assertEquals("Definition of service [service-non-existent] not found", exception.getMessage());
+      Properties result = PgServiceConfParser.getServiceProperties("empty-service1");
       assertNotNull(result);
       assertTrue(result.isEmpty());
     });
@@ -256,7 +256,7 @@ class PgServiceConfParserTest {
   // "PGSYSCONFDIR"                 : exist
   // <service>                      : exist
   @Test
-  void pgService52() throws Exception {
+  public void pgService52() throws Exception {
     URL urlPath = getClass().getResource("/pg_service");
     assertNotNull(urlPath);
     URL urlFileEnv = getClass().getResource("/pg_service/pgservicefileEnv.conf");
@@ -269,10 +269,10 @@ class PgServiceConfParserTest {
     ).execute(() -> {
       Properties result = PgServiceConfParser.getServiceProperties("test-service1");
       assertNotNull(result);
-      assertEquals("test_dbname", PGProperty.DBNAME.get(result));
-      assertEquals("pgservicefileProps-test-host.test.net", PGProperty.HOST.get(result));
-      assertEquals("5433", PGProperty.PORT.get(result));
-      assertEquals("admin", result.get("user"));
+      assertEquals("test_dbname", PGProperty.DBNAME.get2(result));
+      assertEquals("pgservicefileProps-test-host.test.net", PGProperty.HOST.get2(result));
+      assertEquals("5433", PGProperty.PORT.get2(result));
+      assertEquals("admin", PGProperty.USER.get2(result));
       assertEquals(4, result.size());
     });
   }
@@ -283,7 +283,7 @@ class PgServiceConfParserTest {
   // "PGSYSCONFDIR"                 : exist
   // <service>                      : exist
   @Test
-  void pgService53() throws Exception {
+  public void pgService53() throws Exception {
     URL urlPath = getClass().getResource("/pg_service");
     assertNotNull(urlPath);
     URL urlFileEnv = getClass().getResource("/pg_service/pgservicefileEnv.conf");
@@ -293,43 +293,40 @@ class PgServiceConfParserTest {
         new EnvironmentVariables(PGEnvironment.PGSERVICEFILE.getName(), urlFileEnv.getFile(), PGEnvironment.PGSYSCONFDIR.getName(), urlPath.getPath()),
         new SystemProperties(PGEnvironment.ORG_POSTGRESQL_PGSERVICEFILE.getName(), nonExistingFile, "user.home", urlPath.getPath())
     ).execute(() -> {
-      Properties result = PgServiceConfParser.getServiceProperties("test-service1");
-      assertNull(result);
+      JdbcUrlResolverFatalException exception = assertThrows(JdbcUrlResolverFatalException.class, () -> PgServiceConfParser.getServiceProperties("test-service1"));
+      assertEquals("Failed to handle resource [non-existing-file.conf] with error [non-existing-file.conf (No such file or directory)]", exception.getMessage());
     });
   }
 
 
   // resource content read tests
   @Test
-  void pgService61() throws Exception {
+  public void pgService61() throws Exception {
     URL urlPath = getClass().getResource("/pg_service");
     assertNotNull(urlPath);
     Resources.with(
         new EnvironmentVariables(PGEnvironment.PGSERVICEFILE.getName(), "", "APPDATA", urlPath.getPath(), PGEnvironment.PGSYSCONFDIR.getName(), ""),
         new SystemProperties(PGEnvironment.ORG_POSTGRESQL_PGSERVICEFILE.getName(), "", "user.home", urlPath.getPath())
     ).execute(() -> {
-      Properties result;
+      JdbcUrlResolverFatalException exception;
       // fail if there is space between key and equal sign
-      result = PgServiceConfParser.getServiceProperties("fail-case-1");
-      assertNull(result);
+      exception = assertThrows(JdbcUrlResolverFatalException.class, () -> PgServiceConfParser.getServiceProperties("fail-case-1"));
+      assertEquals("Got invalid key: line number [19], value [host =local-somehost1]", exception.getMessage());
       // service name is case-sensitive
-      result = PgServiceConfParser.getServiceProperties("fail-case-2");
-      assertNull(result);
-      // service name is case-sensitive
-      result = PgServiceConfParser.getServiceProperties("fail-case-2");
-      assertNull(result);
+      exception = assertThrows(JdbcUrlResolverFatalException.class, () -> PgServiceConfParser.getServiceProperties("fail-case-2"));
+      assertEquals("Definition of service [fail-case-2] not found", exception.getMessage());
       // invalid line in the section
-      result = PgServiceConfParser.getServiceProperties("fail-case-3");
-      assertNull(result);
+      exception = assertThrows(JdbcUrlResolverFatalException.class, () -> PgServiceConfParser.getServiceProperties("fail-case-3"));
+      assertEquals("Not valid line: line number [27], value [host]", exception.getMessage());
       // service name: space before and after name becomes part of name
-      result = PgServiceConfParser.getServiceProperties(" success-case-3 ");
+      Properties result = PgServiceConfParser.getServiceProperties(" success-case-3 ");
       assertNotNull(result);
-      assertEquals("local-somehost3", PGProperty.HOST.get(result));
+      assertEquals("local-somehost3", PGProperty.HOST.get2(result));
       assertEquals(1, result.size());
       // service name: space inside name is part of name
       result = PgServiceConfParser.getServiceProperties("success case 4");
       assertNotNull(result);
-      assertEquals("local-somehost4", PGProperty.HOST.get(result));
+      assertEquals("local-somehost4", PGProperty.HOST.get2(result));
       assertEquals(1, result.size());
     });
   }